fix compiler warnings
Former-commit-id: c279f8b8d6a139c32bf3f836f30daa29cc38d357

+++ contrib/applications/platform/P300Stimulator/src/properties/ovexP300InterfacePropertyReader.cpp
@@ -64,7 +64,7 @@ void P300InterfacePropertyReader::openChild(const char* sName, const char** sAtt
 				//m_ui32WindowHeight = strtoul(*(sAttributeValue+it),NULL,0);
 				m_f32WindowHeight = static_cast<float32>(m_pKernelContext->getConfigurationManager().expandAsUInteger(CString(*(sAttributeValue+it))));
 			else if(CString(*(sAttributeName+it))==CString("index"))
-				m_ui32MonitorIndex = static_cast<float32>(m_pKernelContext->getConfigurationManager().expandAsUInteger(CString(*(sAttributeValue+it))));
+				m_ui32MonitorIndex = static_cast<uint32>(m_pKernelContext->getConfigurationManager().expandAsUInteger(CString(*(sAttributeValue+it))));
 		}
 	}
 	if (CString(sName)==CString("FeedbackPresentation"))
